requirements.txt
click
PyYAML 